Revert "Add CXL-IDE skid/containment mode support"
This reverts commit e611a6afaa3db12e83337ac0ce3e7af78b6c4a7c.
