+++ teeio-validator/include/cxl_ide.h
@@ -38,8 +38,6 @@ typedef enum {
   CXL_IDE_CONFIGURATION_TYPE_DEFAULT = 0,
   CXL_IDE_CONFIGURATION_TYPE_PCRC,
   CXL_IDE_CONFIGURATION_TYPE_IDE_STOP,
-  CXL_IDE_CONFIGURATION_TYPE_SKID_MODE,
-  CXL_IDE_CONFIGURATION_TYPE_CONTAINMENT_MODE,
   CXL_IDE_CONFIGURATION_TYPE_NUM
 } CXL_IDE_CONFIGURATION_TYPE;
 
@@ -47,9 +45,7 @@ typedef enum {
 #define CXL_LINK_IDE_CONFIGURATION_BITMASK \
   ((CXL_BIT_MASK(CXL_IDE_CONFIGURATION_TYPE_DEFAULT)) | \
   (CXL_BIT_MASK(CXL_IDE_CONFIGURATION_TYPE_PCRC)) | \
-  (CXL_BIT_MASK(CXL_IDE_CONFIGURATION_TYPE_IDE_STOP)) | \
-  (CXL_BIT_MASK(CXL_IDE_CONFIGURATION_TYPE_SKID_MODE)) | \
-  (CXL_BIT_MASK(CXL_IDE_CONFIGURATION_TYPE_CONTAINMENT_MODE)))
+  (CXL_BIT_MASK(CXL_IDE_CONFIGURATION_TYPE_IDE_STOP)))
 
 #pragma pack(1)
