+++ teeio-validator/include/ide_test.h
@@ -394,7 +394,6 @@ typedef struct {
   uint32_t signature;
   ide_common_test_suite_context_t *suite_context;
   IDE_TEST_TOPOLOGY *top;
-  int config_id;
 
   ide_common_test_port_context_t root_port;
   ide_common_test_port_context_t upper_port;
@@ -410,7 +409,6 @@ typedef struct {
   uint32_t signature;
   ide_common_test_suite_context_t *suite_context;
   IDE_TEST_TOPOLOGY *top;
-  int config_id;
 
   ide_common_test_port_context_t root_port;
   ide_common_test_port_context_t upper_port;
@@ -434,7 +432,6 @@ typedef struct {
   uint32_t signature;
   ide_common_test_suite_context_t *suite_context;
   IDE_TEST_TOPOLOGY *top;
-  int config_id;
 
   ide_common_test_port_context_t root_port;
   ide_common_test_port_context_t upper_port;
@@ -569,7 +566,6 @@ typedef ide_test_group_funcs_t*(*teeio_get_test_group_funcs_func_t) (int top_typ
 typedef ide_test_case_funcs_t*(*teeio_get_test_case_funcs_func_t) (int case_class, int case_id);
 typedef ide_test_case_name_t*(*teeio_get_test_case_name_func_t) (int case_class);
 typedef void*(*teeio_alloc_test_group_context_func_t)(void);
-typedef bool(*teeio_check_configuration_bitmap_func_t) (uint32_t* bitmask);
 
 typedef struct {
   teeio_get_test_configuration_name_func_t get_configuration_name_func;
@@ -579,7 +575,6 @@ typedef struct {
   teeio_get_test_case_funcs_func_t get_case_funcs_func;
   teeio_get_test_case_name_func_t get_case_name_func;
   teeio_alloc_test_group_context_func_t alloc_test_group_context_func;
-  teeio_check_configuration_bitmap_func_t check_configuration_bitmap_func;
 } teeio_test_funcs_t;
 
 typedef struct _ide_run_test_suite ide_run_test_suite_t;
