+++ teeio-validator/include/intel_keyp.h
@@ -23,7 +23,7 @@ typedef enum {
 
 typedef enum {
   INTEL_CXL_IDE_MODE_RSVD = 0,
-  INTEL_CXL_IDE_MODE_CONTAINMENT,
+  INTEL_CXL_IDE_MODE_DETER_CONTAINMENT,
   INTEL_CXL_IDE_MODE_SKID,
   INTEL_CXL_IDE_MODE_MAX
 } INTEL_CXL_IDE_MODE;
